backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
itsdangerous==2.1.2
psycopg2-binary==2.9.9
